refactor(config): extract comment_writer to a single shared module
The gateway PATCH/PUT handlers and the CLI `config set --comment` /
`config patch` flow had two near-identical implementations of TOML
key-decor walking + comment-prefix construction. Move both into
`zeroclaw_config::comment_writer` as the single source of truth.

Public surface:
- `apply_comments(config_path, &[(path, comment)])` — apply a batch.
- `decorate_key(table, dotted, comment)` — single-leaf operation.
- `build_comment_prefix(existing, comment)` — pure prefix builder
  (kept public so callers writing into a doc they already own can
  share the same prefix-clearing semantics).

Gateway `api_config.rs` and CLI `main.rs` now both delegate to the
shared helper. The unit tests for `build_comment_prefix` move to
`comment_writer::tests` (gateway test module gets a one-liner pointer
so future readers know where the shared coverage lives) and gain two
new edge cases: multi-blank-line preservation and multi-line comment
input.

Refs #6175. Adversarial review pass 1, finding P1.1.
