+++ crates/zeroclaw-config/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod api_error;
 pub mod autonomy;
+pub mod comment_writer;
 pub mod cost;
 pub mod domain_matcher;
 pub mod helpers;
