+++ crates/zeroclaw-gateway/src/api_config.rs
@@ -372,97 +372,6 @@ async fn persist_and_swap(
     Ok(())
 }
 
-/// Decorate the on-disk TOML file with comments captured from PATCH/PUT ops.
-///
-/// Called after `Config::save()` (which already preserves existing comments
-/// via `migration::sync_table`). For each `(path, comment)` pair, walks the
-/// toml_edit document to the target leaf and prepends `# {comment}\n` as
-/// the leading decoration. An empty comment string clears any existing
-/// `# `-prefixed comment lines from the leaf's leading decor (other
-/// whitespace and blank lines are left intact).
-///
-/// Best-effort: silently skips paths that don't resolve to a leaf value.
-/// Failure to read or write the file leaves the surface unchanged.
-async fn apply_comments(
-    config_path: &std::path::Path,
-    annotations: &[(String, String)],
-) -> Result<(), std::io::Error> {
-    if annotations.is_empty() {
-        return Ok(());
-    }
-
-    let raw = tokio::fs::read_to_string(config_path).await?;
-    let mut doc: toml_edit::DocumentMut = match raw.parse() {
-        Ok(d) => d,
-        Err(_) => return Ok(()), // unparseable; bail without touching the file
-    };
-
-    for (path, comment) in annotations {
-        decorate_key(doc.as_table_mut(), path, comment);
-    }
-
-    tokio::fs::write(config_path, doc.to_string()).await?;
-    Ok(())
-}
-
-/// Walk to the leaf key for `dotted` and decorate it with `# {comment}\n`,
-/// preserving any non-comment whitespace already in the prefix. Empty comment
-/// strips comment lines from the existing prefix.
-fn decorate_key(root: &mut toml_edit::Table, dotted: &str, comment: &str) {
-    let segments: Vec<&str> = dotted.split('.').collect();
-    let (last, rest) = match segments.split_last() {
-        Some(s) => s,
-        None => return,
-    };
-    fn walk<'a>(
-        table: &'a mut toml_edit::Table,
-        segs: &[&str],
-    ) -> Option<&'a mut toml_edit::Table> {
-        let mut cursor = table;
-        for seg in segs {
-            cursor = cursor.get_mut(seg)?.as_table_mut()?;
-        }
-        Some(cursor)
-    }
-    let table = match walk(root, rest) {
-        Some(t) => t,
-        None => return,
-    };
-    if let Some(mut key) = table.key_mut(last) {
-        let decor = key.leaf_decor_mut();
-        let new_prefix = build_comment_prefix(decor.prefix(), comment);
-        decor.set_prefix(new_prefix);
-    }
-}
-
-/// Build the new leading decor for a leaf, applying the `# {comment}\n` line
-/// while preserving any blank-line whitespace that preceded it. When the
-/// comment is empty, strips comment lines from the existing prefix.
-fn build_comment_prefix(existing: Option<&toml_edit::RawString>, comment: &str) -> String {
-    let prev = existing.and_then(|r| r.as_str()).unwrap_or("");
-
-    // Split existing prefix into non-comment whitespace lines (kept) and
-    // comment lines (replaced).
-    let mut kept: Vec<&str> = Vec::new();
-    for line in prev.split_inclusive('\n') {
-        let trimmed = line.trim_start();
-        if trimmed.starts_with('#') {
-            continue;
-        }
-        kept.push(line);
-    }
-
-    let mut out: String = kept.join("");
-    if !comment.is_empty() {
-        for line in comment.lines() {
-            out.push_str("# ");
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    out
-}
-
 /// Compute drift between the in-memory config and what's on disk right now.
 /// Returns one entry per drifted property; empty when in-memory and disk
 /// agree (or when the on-disk file can't be parsed).
@@ -639,7 +548,7 @@ pub async fn handle_prop_put(
     }
     if let Some(comment) = body.comment.as_ref() {
         let annotations = [(body.path.clone(), comment.clone())];
-        if let Err(e) = apply_comments(&config_path, &annotations).await {
+        if let Err(e) = zeroclaw_config::comment_writer::apply_comments(&config_path, &annotations).await {
             tracing::warn!(error = %e, "failed to apply PUT comment to config.toml");
         }
     }
@@ -947,7 +856,7 @@ pub async fn handle_patch(
         return error_response(e);
     }
     if !annotations.is_empty()
-        && let Err(e) = apply_comments(&config_path, &annotations).await
+        && let Err(e) = zeroclaw_config::comment_writer::apply_comments(&config_path, &annotations).await
     {
         // Comments are best-effort decoration; surface as a non-fatal warn.
         // The patch itself succeeded — return success but log the failure.
@@ -1282,31 +1191,8 @@ mod tests {
         assert!(result.is_err());
     }
 
-    #[test]
-    fn build_comment_prefix_appends_comment_to_blank_prefix() {
-        let out = build_comment_prefix(None, "set during onboarding");
-        assert_eq!(out, "# set during onboarding\n");
-    }
-
-    #[test]
-    fn build_comment_prefix_replaces_existing_comment_lines() {
-        // Simulate a doc that already has a comment + a blank line.
-        let raw = toml_edit::RawString::from("\n# old reason\n");
-        let out = build_comment_prefix(Some(&raw), "new reason");
-        assert!(out.contains("# new reason\n"));
-        assert!(!out.contains("old reason"));
-        // Blank line preserved.
-        assert!(out.starts_with('\n'));
-    }
-
-    #[test]
-    fn build_comment_prefix_empty_comment_strips_existing() {
-        let raw = toml_edit::RawString::from("\n# stale\n");
-        let out = build_comment_prefix(Some(&raw), "");
-        assert!(!out.contains('#'));
-        // Blank line preserved.
-        assert_eq!(out, "\n");
-    }
+    // build_comment_prefix tests live in zeroclaw_config::comment_writer
+    // — shared helper, single source of truth.
 
     #[test]
     fn json_to_setprop_string_accepts_bool_string_for_bool_field() {
@@ -1443,7 +1329,7 @@ mod tests {
         cfg.set_prop("gateway.host", "10.0.0.5").expect("set_prop");
         cfg.save().await.expect("save");
 
-        apply_comments(
+        zeroclaw_config::comment_writer::apply_comments(
             &path,
             &[("gateway.host".into(), "raised after Q3 backlog".into())],
         )
@@ -1535,10 +1421,10 @@ mod tests {
         cfg.set_prop("gateway.host", "10.0.0.5").expect("set_prop");
         cfg.save().await.expect("save");
 
-        apply_comments(&path, &[("gateway.host".into(), "first reason".into())])
+        zeroclaw_config::comment_writer::apply_comments(&path, &[("gateway.host".into(), "first reason".into())])
             .await
             .expect("apply first comment");
-        apply_comments(&path, &[("gateway.host".into(), String::new())])
+        zeroclaw_config::comment_writer::apply_comments(&path, &[("gateway.host".into(), String::new())])
             .await
             .expect("apply empty");
